Refactor dashboard load
