+++ src/index.html
@@ -55,7 +55,7 @@
     </article>
 
     <!-- DASHBOARD CONTAINER -->
-    <article id="dashboardContainer">
+    <article id="dashboardContainer" class="hidden">
         <header>
             <!-- LOGO -->
             <span>
