+++ src/scripts/main.js
@@ -1,28 +1,21 @@
-import API from './data.js'
 import userRegistration from './registration/registration.js'
-import renderToDom from "./render.js"
+import loadDashboard from './dashboard.js';
+
+const registrationContainer = document.querySelector("#registrationContainer")
+const hiddenDashboard = document.querySelector("#dashboardContainer")
+
+if (sessionStorage.getItem("currentUser")) {
+    loadDashboard()
+}
 
 userRegistration.clickRegistrationLink();
 userRegistration.registrationFormValidator();
 userRegistration.registeredUserLogin();
 
-
-renderToDom.tasksList()
-renderToDom.eventsList()
-renderToDom.articlesList()
-
-
-const registrationContainer= document.querySelector("#registrationContainer")
-const hiddenDashboard = document.querySelector("#dashboardContainer")
-     hiddenDashboard.style.display = "none"
-
-//place in main js
-   //user requested to log out, reset DOM to registration page
-   //STILL REQUIRED clear user session from browser.
+// Logout button
 const userLogOutRequest = document.querySelector(".logOutButton")
-userLogOutRequest.addEventListener("click", event => { 
-    hiddenDashboard.style.display = "none"
-    registrationContainer.style.display = "block"
-
-    
+userLogOutRequest.addEventListener("click", event => {
+    registrationContainer.classList.toggle("hidden")
+    hiddenDashboard.classList.toggle("hidden")
+    sessionStorage.clear()
 })
