+++ src/scripts/registration/registration.js
@@ -1,6 +1,6 @@
 import API from '../data.js'
 import createUserObject from './createUser.js'
-import renderToDom from "../render.js"
+import loadDashboard from '../dashboard.js'
 
 const userRegistration = {
     //registered user login 
@@ -10,8 +10,6 @@ const userRegistration = {
         const passwordLogin = document.querySelector("#loginPassword")
         const signInButton = document.querySelector("#signInButton")
         const loginInputFields = document.getElementsByClassName("login__input")
-        const registrationContainer= document.querySelector("#registrationContainer")
-        const hiddenDashboard = document.querySelector("#dashboardContainer")
        
         //loop through class names for inputs and not allowing sign-in button to be clicked until all fields are filled out
         for (let i = 0; i < loginInputFields.length; i++) {
@@ -47,9 +45,7 @@ const userRegistration = {
                         passwordLogin.style.borderColor = ""
                         console.log("stored user:", sessionStorage.getItem("currentUser"))
                         //***need to display main dashboard****
-                        registrationContainer.style.display = "none"
-                        hiddenDashboard.style.display = "block"
-                        API.getMessagesData().then(messagesCollection => renderToDom.messagesList(messagesCollection))
+                        loadDashboard()
                         //if email not found, alert user
                     } else if (!findEmail) {
                         alert("Your email address does not match existing user");
@@ -171,9 +167,9 @@ const userRegistration = {
                             console.log("stored userId:", sessionStorage.getItem("currentUser"))
                             this.clearRegistrationFields();
                             registerButton.disabled = true;
-                            hiddenDashboard.style.display = "block"
-                            registrationContainer.style.display = "none"
-                            API.getMessagesData().then(messagesCollection => renderToDom.messagesList(messagesCollection))
+                            
+                            // Load the dashboard
+                            loadDashboard()
                         })
                     }
                 })
